+++ test/converters/html-converter.js
@@ -1,23 +1,22 @@
 require('chai').should();
 const isHTML = require('is-html');
-const HtmlConverter = require('../../lib/converters/htmlConverter');
-const path = require('path');
+const HtmlConverter = require('../../lib/converters/html-converter');
 
-const EX_1_HTML_INFO = require('../resources/exampleData').EX_1_HTML_CONVERTER_HTML_INFO;
-const EX_1_DATA = require('../resources/exampleData').EX_1_HTML_CONVERTER_DATA;
-const EX_2_HTML_INFO = require('../resources/exampleData').EX_2_HTML_CONVERTER_HTML_INFO;
-const EX_3_HTML_INFO = require('../resources/exampleData').EX_3_HTML_CONVERTER_HTML_INFO;
-const EX_3_DATA = require('../resources/exampleData').EX_3_HTML_CONVERTER_DATA;
-const EX_3_OUTPUT = require('../resources/exampleData').EX_3_HTML_CONVERTER_OUTPUT;
-const EX_4_HTML_INFO = require('../resources/exampleData').EX_4_HTML_CONVERTER_HTML_INFO;
-const EX_4_OUTPUT = require('../resources/exampleData').EX_4_HTML_CONVERTER_OUTPUT;
+const EX_1_HTML_INFO = require('../resources/example-data').EX_1_HTML_CONVERTER_HTML_INFO;
+const EX_1_DATA = require('../resources/example-data').EX_1_HTML_CONVERTER_DATA;
+const EX_2_HTML_INFO = require('../resources/example-data').EX_2_HTML_CONVERTER_HTML_INFO;
+const EX_3_HTML_INFO = require('../resources/example-data').EX_3_HTML_CONVERTER_HTML_INFO;
+const EX_3_DATA = require('../resources/example-data').EX_3_HTML_CONVERTER_DATA;
+const EX_3_OUTPUT = require('../resources/example-data').EX_3_HTML_CONVERTER_OUTPUT;
+const EX_4_HTML_INFO = require('../resources/example-data').EX_4_HTML_CONVERTER_HTML_INFO;
+const EX_4_OUTPUT = require('../resources/example-data').EX_4_HTML_CONVERTER_OUTPUT;
 
 describe('HtmlConverter', function () {
 
   before(function () {
   });
 
-  describe('#functionality()', function () {
+  describe('# Functionality', function () {
     it('should be able to convert the given JSON to HTML using the given template and engine', async () => {
       const html = await HtmlConverter.convert(EX_1_HTML_INFO, EX_1_DATA);
       isHTML(html).should.be.true;