+++ test/parsers/data-sources-parser.js
@@ -1,15 +1,15 @@
 require('chai').should();
 
-const parseDataSources = require('../../lib/parsers/dataSourcesParser');
-const GraphQLLDHandler = require('../../lib/handlers/graphQLLDHandler');
-const createLogger = require('../../lib/createLogger');
+const parseDataSources = require('../../lib/parsers/data-sources-parser');
+const GraphQLLDHandler = require('../../lib/handlers/graphql-ld-handler');
+const createLogger = require('../../lib/create-logger');
 const path = require('path');
 const YAML = require('yaml');
 const fs = require('fs');
 
 describe('DataSourcesParser', function () {
 
-  describe('#query', function () {
+  describe('# Query', function () {
     it('should be able to execute query to get list of data sources', async function () {
       const file = fs.readFileSync(path.resolve(__dirname, '../resources/query-datasources/single-query/config.yaml'), 'utf8');
       const pipeModulesPath = path.resolve(__dirname, '../resources/query-datasources/single-query');