+++ test/parsers/route-parser.js
@@ -2,9 +2,9 @@ require('chai').should();
 
 const fs = require('fs');
 const path = require('path');
-const parseRoute = require('../../lib/parsers/routeParser');
+const parseRoute = require('../../lib/parsers/route-parser');
 
-const CONFIG_FILE = '../resources/config_test_example.yaml';
+const CONFIG_FILE = '../resources/config.yaml';
 
 describe('RouteParser', function () {
   {
@@ -14,7 +14,7 @@ describe('RouteParser', function () {
       this.output = parseRoute('/movies/{actor}', 'get');
     });
 
-    describe('#functionality()', function () {
+    describe('# Functionality', function () {
       it('should be able to parse, extract and format route information correctly from a YAML config file', function () {
         this.output.should.eql(
           {
@@ -25,7 +25,7 @@ describe('RouteParser', function () {
       });
     });
 
-    describe('#outputFormat()', function () {
+    describe('# Output format', function () {
       it('output object should have {path, method} properties', function () {
         this.output.should.have.property('path');
         this.output.should.have.property('method');